Share index.html between all web-samples
Since most of the web-samples initialize themselves the same way, there is no reason to maintain individual copies of the file for each sample. The original “spinning disk” -sample has a slightly different DOM-structure, so … for the time being, it can still have its own index.html (though hopefully I can get rid of that eventually) … aside from the slight differences, the basic functionality of that sample is also identical to the others.